[⚙️ chore] EmptyView 컴포넌트 중 텍스트 상수 파일 분리 #14
#17 변경사항 적용
