+++ src/components/common/EmptyView.tsx
@@ -1,29 +1,11 @@
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 import Image from 'next/image';
+import emptyText from '@/constants/emptyText';
 
 interface IEmptyView {
   pageType: '마이챌린지' | '챌린지후기' | '내후기' | '예치금내역';
 }
-
-const emptyText = {
-  마이챌린지: [
-    '아직 내 챌린지가 없어요',
-    '챌린지를 신청하고 내 챌린지를 시작해보세요!',
-  ],
-  챌린지후기: [
-    '남기신 후기가 없어요',
-    '챌린지에 참여하시고 후기를 남겨보세요!',
-  ],
-  내후기: [
-    '남기신 후기가 없어요',
-    '마이챌린지에서 참여하신 챌린지 후기를 남겨보세요!',
-  ],
-  예치금내역: [
-    '예치금 내역이 없어요',
-    '예치금을 걸어서 챌린지에 참여해보세요!',
-  ],
-};
 
 export default function EmptyView({ pageType }: IEmptyView) {
   const mainText = emptyText[pageType][0];
